API: Move semantic generics into `marker_api::sem::generic`
