+++ marker_api/src/sem/ty.rs
@@ -64,7 +64,7 @@ pub enum SemTyKind<'ast> {
     // ================================
     // User defined types
     // ================================
-    /// A user defined data type, identified by an [`TyDefId`](marker_api::ast::TyDefId)
+    /// A user defined data type, identified by an [`TyDefId`](crate::ast::TyDefId)
     Adt(&'ast SemAdtTy<'ast>),
     /// A generic type defined by a generic parameter
     Generic(&'ast SemGenericTy<'ast>),
